Sets up Pac-Man AI with 2 algorithms
Using the Pac-Man Java project, breadth first search and depth first search are implemented to control the AI of Ms. Pac-Man. In the implementation of these search algorithms, the game state is duplicated and evaluated using the game score and total lives remaining, and the best game state is determined using the alotted time. Using the best game state, only the initial move concerning the path is chosen as the algorithm will be executed every time Ms. Pac-Man moves 1 space.