Adds/Revises Tree Search Algorithms
Breadth first search, depth first search, iterative deepening, and A* search are either revised or added to the list of possible algorithms to be used to control the AI of Pac-Man. In addition, the evaluation function is rehauled to take into account the current level, the number of lives remaining, the nearest ghost, the nearest power pill, the nearest regular pill, the number of power pills, the number of regular pills, and the in-game score with different weights on all of them. The highest weights are given to the level and the number of lives remaining. The next highest is the distance away from the nearest ghost that is not edible.

In addition, the aggressive ghost AI is now being used instead of the starter one, as the starter AI seems to be glitched to a certain extent.