Adds the kNN algorithm
In order to implement the kNN algorithm, a file containing moves of a previous Pac-Man game must be used ("replay.txt"). This data is read in by the Pac-Man AI, and the kNN algorithm uses it to determine how to act depending on the distance between Pac-Man and the enemy ghosts, and the distance between Pac-Man and the nearest active pill. It checks the data using the "replay" and find the k occurrences where a simular situation occurred (similar distances) and acts either with the mindset of escaping from the ghosts, or taking the nearest pill.