Implement decision tree algorithm
The decision tree algorithm utilizes data from the "replay.txt" file, and uses two main conditions to determine the best possible move for Pac-Man. First it looks at all of the possible moves for Pac-Man (Up, Left, Right, Down), and then it looks at the last move of the nearest Ghost. It looks into the history and determines which move it should take in this exact situation.