Adds Monte-Carlo BFS algorithm
The algorithm is a mix between the regular Monte-Carlo search method and breadthe-first search. Rather than playing out entire scenarios with select moves, all of the moves are expanded equally until there is no time left. All of the expansions are evaluated and determined to either be good or bad moves which are used to evaluate ratios for each initial move. The initial move with the highest ratio is then chosen and acted upon. Also unlike the Monte-Carlo search method, the non-leaf nodes are judged using the evaluation function as well.